Replace off_t with size_t where unsigned value is required
This includes the offset stored within the kernel "file" structure
and the off_t with the VFS functions

+++ kernel/src/driver/ramdisk.c
@@ -64,7 +64,7 @@ static ssize_t ramdisk_write_block(void *buf, size_t blocknum)
 
 static char blk[BLOCK_SIZE];
 
-ssize_t ramdisk_read(void *buf, size_t size, off_t off)
+ssize_t ramdisk_read(void *buf, size_t size, size_t off)
 {
     unsigned int nblk;
     ssize_t left;
@@ -98,7 +98,7 @@ ssize_t ramdisk_read(void *buf, size_t size, off_t off)
     return size-left;
 }
 
-ssize_t ramdisk_write(const void *buf, size_t size, off_t off)
+ssize_t ramdisk_write(const void *buf, size_t size, size_t off)
 {
     return -1; /* TODO */
 }
